requirements.txt
fastapi
uvicorn[standard]
pillow
celery  # Celery와 Redis 또는 RabbitMQ 지원
